+++ src/main/resources/reference.conf
@@ -6,11 +6,4 @@ kamon {
   netty {
 
   }
-
-  modules {
-    kamon-netty {
-      requires-aspectj = yes
-    }
-  }
-
 }
