+++ src/main/scala/kamon/netty/instrumentation/NioEventLoopInstrumentation.scala
@@ -23,7 +23,6 @@ import io.netty.channel.{ChannelFuture, ChannelFutureListener}
 import kamon.metric.Gauge
 import kamon.netty.Metrics
 import kamon.netty.instrumentation.EventLoopUtils._
-import kamon.netty.util.MonitoredQueue
 import org.aspectj.lang.ProceedingJoinPoint
 import org.aspectj.lang.annotation._
